OC-27826: Move OC-specific JS utils out of upstream files into jsapp/js/oc/
- Add jsapp/js/oc/permissionUtils.ts with userWithSameSubdomainAsAssetOwner
- Add jsapp/js/oc/utils.ts (renamed from jsapp/js/ocutils.ts)
- Remove userWithSameSubdomainAsAssetOwner from upstream permissions/utils.ts
- Update all importers (session.ts, myLibraryStore.ts, libraryTypeFilterStore.ts,
  permProtectedRoute.tsx, main.js, view.row.coffee) to use #/oc/utils or #/oc/permissionUtils
- Delete jsapp/js/ocutils.ts
