+++ jsapp/js/router/permProtectedRoute.tsx
@@ -4,9 +4,10 @@ import { actions } from '#/actions'
 import assetStore from '#/assetStore'
 import LoadingSpinner from '#/components/common/loadingSpinner'
 import type { PermissionCodename } from '#/components/permissions/permConstants'
-import { userCan, userCanPartially, userWithSameSubdomainAsAssetOwner } from '#/components/permissions/utils'
+import { userCan, userCanPartially } from '#/components/permissions/utils'
 import { decodeURLParamWithSlash } from '#/components/processing/routes.utils'
 import type { AssetResponse, FailResponse } from '#/dataInterface'
+import { userWithSameSubdomainAsAssetOwner } from '#/oc/permissionUtils'
 import AccessDenied from '#/router/accessDenied'
 import type { WithRouterProps } from '#/router/legacy'
 import { withRouter } from './legacy'
